agent: extend util with health util and make erc20_abi singleton
